Fixed zipper tests
Previous test helper never actually compared the two zips, instead
only comparing the expected zip to itself. This seems to have been
a copy-paste mistake.

New code checks that the zips have the same amount of entries and
that they have the same names. Checking for size etc. is not done
because the commons-compress does not have good methods for this.