Add test cases
Added test case directory structure

This commit outlines the required test cases and their directory
structure.

Deleted redundant test project

This commit removes the redundant tmc_test_project directory as it's
contents will not be used in test case development

Added nolib test cases

Added test cases for passing/failing projects in which no libraries are
included

Added test cases

Added test case with tests passing for project with app using a single
library.

Added a test case for failing tests on app using a single library.

Added test case failing compile with single lib that compiles

This commit adds the test case in which the library compiles as intended
however the app fails to compile. Also fixes test_case_test_runner.cpp
files already pushed to be compatible with unix environment (Qtest->QTest)

Added the actual files described in the previous commit

I was in the wrong directory when adding committing and pushing the
earlier commit ;-DAdded the actual files described in the previous commit

I was in the wrong directory when adding committing and pushing the
earlier commit ;-D

Added test case for library not compiling

Added test case in which the included library fails to compile

Added qInfo with points to all test cases

Added relevant qInfo fields with relevant output to all test case
test_case_test_runner/test_case_test_runner.cpp files

Added test case for passing project with multiple libraries

Added the test case for a project with multiple (in our case two)
libraries for which all tests pass and which builds correctly.

Added test case for single lib multiple tests for one function

Test case with single library where two tests for a single function with
the question on how do we assign scores in this kind of cases?

Added failing single lib same point

Added test case in which there are two tests for one function, of which
one passes and one fails, and the assumption is that these two tests are
required to recieve a point. So if one test fails and one passes you will
still not get a point because they both have to be right. There is also
one unrelated passing test for another function.

Added test case for passing nolib same point

Added a test case in which there are two tests for one function with all
test passing and no libraries in use

Added test case for failing nolib same point

Added a test case in which the project does not use a library, two tests
(of which one fails) tests the second function. One passing test tests the
first function.

Added test case for passing multiple lib same point

Added a test case in which the tested functions are in separate libraries
to test that multiple libraries compile right. The function in the test
library 2 is tested by tests which both pass.